Emulator runs more smoothly.
The call to setTimeout() for batching instructions has
been moved to the start of the function so that it will
run another instance almost immediately after the first
one exits, instead of waiting for a ~10ms gap between
batches of instructions.

Additionally, there is not a fixed amount of instructions
per batch. The emulator stops executing instructions when
it is at or past when the next batch should begin, allowing
it to adapt to different hardware.